+++ src/main/java/fr/eni/tp/encheres/dal/ArticleDAO.java
@@ -5,10 +5,10 @@ import java.util.List;
 import fr.eni.tp.encheres.bo.Article;
 
 public interface ArticleDAO {
-	Article read(long articleId);
+	Article read(int articleId);
 	List<Article> findAll();
-	List<Article> findByCategory(long categoryId);
+	List<Article> findByCategory(int categoryId);
 	List<Article> findByName(String name);
 	void create(Article article);
-	void delete(long articleId);
+	void delete(int articleId);
 }
